Revert change and stop running remove duplicates.
Revert "Don't merge types of resources"

This reverts commit 6c9600820e166084c7b214376aba7a55bea390e0, but leaves
the tests that were added.  Added new test. These test are the so that,
if someone tries the same change I made, they will see the test that
they need to handle.

Don't run remove duplicates in -O and -Os

Romve duplicates was run to help reduce compile time when looking for
types in the type manager.  I've run compile time test on three sets
of shaders, and the compile time does not seem to change.

It should be safe to remove it.